Move graphql resolvers (those that were already nicely separated) into server directory
