+++ packages/lesswrong/lib/collections/comments/index.js
@@ -1,7 +1,6 @@
 export * from './collection.js';
 
 import './custom_fields.js';
-import './graphql.js'
 import './helpers.js';
 import './permissions.js';
 import './views.js';
